task: move unused codebase to the new package
