+++ auth/code/device_flow_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package auth
+package code
 
 import (
 	"fmt"
